+++ src/util/compile.rs
@@ -1,10 +1,10 @@
 //! Compile options for the compiler.
 
-use std::sync::Mutex;
+use std::{ops::RangeInclusive, sync::Mutex};
 
 use getset::Getters;
 
-use crate::datapack::Function;
+use crate::{datapack::Function, prelude::Datapack};
 
 use super::extendable_queue::ExtendableQueue;
 
@@ -14,12 +14,34 @@ use super::extendable_queue::ExtendableQueue;
 #[derive(Debug, Clone)]
 pub struct CompileOptions {
     /// Whether to compile in debug mode.
-    pub debug: bool,
+    pub(crate) debug: bool,
+
+    pub(crate) pack_formats: RangeInclusive<u8>,
+}
+
+impl CompileOptions {
+    /// Set whether to compile in debug mode.
+    #[must_use]
+    pub fn with_debug(self, debug: bool) -> Self {
+        Self { debug, ..self }
+    }
+
+    /// Set the pack format of the datapack.
+    #[must_use]
+    pub fn with_pack_formats(self, pack_formats: RangeInclusive<u8>) -> Self {
+        Self {
+            pack_formats,
+            ..self
+        }
+    }
 }
 
 impl Default for CompileOptions {
     fn default() -> Self {
-        Self { debug: true }
+        Self {
+            debug: true,
+            pack_formats: Datapack::LATEST_FORMAT..=Datapack::LATEST_FORMAT,
+        }
     }
 }