+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Execute
     - Debug
     - Group
+- Validate function for checking pack format compatibility
 - Virtual file system
 
 ### Changed